api delete account + validare date intrare + alte schimbari minore

+++ api/category/delete.php
@@ -26,6 +26,12 @@ $data = json_decode(file_get_contents("php://input"));
 
 $category->id_categorie = $data->id_categorie;
 
+if(!ctype_digit($category->id_categorie)) {
+  http_response_code(400);
+  echo json_encode(array("message" => "Invalid category id detected."));
+  die();
+}
+
 if($category->delete()){
     http_response_code(204);
     echo json_encode(array("message" => "Category was deleted."));
